// nextjs/utils/api/client/models/MenuItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trillion Trees Campaign
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MenuItem
 */
export interface MenuItem {
    /**
     * 
     * @type {number}
     * @memberof MenuItem
     */
    enabled: number;
    /**
     * 
     * @type {number}
     * @memberof MenuItem
     */
    sequence: number;
    /**
     * 
     * @type {string}
     * @memberof MenuItem
     */
    caption: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItem
     */
    help: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItem
     */
    icon: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItem
     */
    uri: string;
}

export function MenuItemFromJSON(json: any): MenuItem {
    return MenuItemFromJSONTyped(json, false);
}

export function MenuItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): MenuItem {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'enabled': json['enabled'],
        'sequence': json['sequence'],
        'caption': json['caption'],
        'help': json['help'],
        'icon': json['icon'],
        'uri': json['uri'],
    };
}

export function MenuItemToJSON(value?: MenuItem | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'enabled': value.enabled,
        'sequence': value.sequence,
        'caption': value.caption,
        'help': value.help,
        'icon': value.icon,
        'uri': value.uri,
    };
}